* FIX: Revert IReader to working state * FIX: Revert CActor::create_entity_condition to working state, and plz, don't change that function anymore

+++ code/engine.vc2008/xrCore/FS.h
@@ -335,7 +335,7 @@ class XRCORE_API IReader : public IReaderBase<IReader>
 protected:
     char* data;
     int Pos;
-    size_t Size;
+    int Size;
     int iterpos;
 
 public:
@@ -359,13 +359,13 @@ protected:
     size_t advance_term_string();
 
 public:
-    size_t elapsed() const { return Size - Pos; }
+    int elapsed() const { return Size - Pos; }
     int tell() const { return Pos; }
     void seek(const int ptr) {
         Pos = ptr;
         VERIFY((Pos <= Size) && (Pos >= 0));
     }
-    size_t length() const { return Size; }
+    int length() const { return Size; }
     void* pointer() const { return &(data[Pos]); }
     void advance(const int cnt) {
         Pos += cnt;
